Fix SSH connection termination, error clearing, and add reset_client.
The SSHClient class wasn't closing except on successful connection. It
appeared this was the proper behaviour as per the documentation, but
changing this to unconditionally close, which gracefully handles resetting
the connection.

self.clear_error() was improperly clearing errors, causing the script to
continually think the SSH timeout error was persisting when it wasn't.

self.reset_client() was added to provide client reset capabilities.